feat: automate benchmark on PR requests using Github Actions (#1037)
This PR adds Github actions to automate benchmark running on PR requests

It will create a [Github Job
Summary](https://github.blog/news-insights/product-news/supercharging-github-actions-with-job-summaries/)
of the benchmark and add it to the benchmark action (see
https://github.com/moscajs/aedes/actions/runs/16239498783#summary-45853979456
)

The benchmark steps run serialized in the hope that performance on a
single runner does not vary too much.
I have been building testing this using
https://github.com/seriousme/aedes/pull/1 which also shows how it
operates.

It also shows that variation between results is inevitable, even when
running on the same runner and the Aedes code is identical !!

In https://github.com/seriousme/aedes/pull/1 I was able to add the
comment automatically to the PR using
`.github/actions/sticky-pr-comment/action.yml` however if you send a PR
from a fork the GITHUB_TOKEN that runs the actions by default (event:
[pull_request](https://docs.github.com/en/actions/reference/events-that-trigger-workflows#pull_request)
) only has `read` access on the PR for security reasons (see [Pull
Request events for forked
repositories](https://docs.github.com/en/actions/reference/events-that-trigger-workflows#pull-request-events-for-forked-repositories))

This can be fixed by attaching to the event:
[pull_request_target](https://docs.github.com/en/actions/reference/events-that-trigger-workflows#pull_request_target),
however Github notes explicitly: *Avoid using this event if you need to
build or run code from the pull request.*

So for now I used Github Job Summary instead.

Kind regards,
Hans

---------

Co-authored-by: Daniel Lando <daniel.sorridi@gmail.com>