chore(dev): fix incorrect automated benchmark threshold handling (#1046)
A bug in the benchmarking report triggered a fail if the benchmark of
the branch to merge scored more than the threshold over the main branch
(ie. the merge branch is better), where it should fail if the score is
more than the threshold below the main branch.

This PR aims to fix that, and also improves reporting in case of a fail.

Kind regards,
Hans